fix: moved `function_references` from ModuleData to ResolveReferences to better suit the mind model
