+++ tests/library_analyzer/processing/api/purity_analysis/__init__.py
@@ -1,6 +1,1 @@
-from tests.library_analyzer.processing.api.purity_analysis.test_get_module_data import (
-    transform_member_access,
-)
-
-__all__ = ["transform_member_access"]
